requirements.txt
google-adk
yfinance
python-dotenv
pydantic
python-math